# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/getMeetingInfo": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "全量获取连麦信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "全量获取连麦信息",
                "operationId": "GetMeetingInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.GetMeetingInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.MeetingInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
        "/joinMeeting": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "成员加入连麦",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "成员加入连麦",
                "operationId": "JoinMeeting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MeetingActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.MeetingInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
        "/leaveMeeting": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "成员下麦",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "成员下麦",
                "operationId": "LeaveMeeting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MeetingActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.MeetingInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
...
        "/updateMeeting": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新成员连麦信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "更新成员连麦信息",
                "operationId": "UpdateMeeting",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MeetingActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.MeetingInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        },
        "/updateMeetingInfo": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "全量更新直播连麦人员信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "全量更新直播连麦人员信息",
                "operationId": "UpdateMeetingInfo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer your-token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.MeetingActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.MeetingInfo"
                        }
                    },
                    "400": {
                        "description": "4xx, 客户端错误",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    },
                    "500": {
                        "description": "5xx, 请求失败",
                        "schema": {
                            "$ref": "#/definitions/models.Status"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.GetMeetingInfoRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "直播Id"
                }
            }
        },
...
        "handler.MeetingActionRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "直播Id"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MeetingMember"
                    }
                }
            }
        },
...
        "models.MeetingInfo": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MeetingMember"
                    }
                }
            }
        },
        "models.MeetingMember": {
            "type": "object",
            "properties": {
                "camera_opened": {
                    "description": "摄像头状态",
                    "type": "boolean"
                },
                "meeting_id": {
                    "description": "会议(连麦)Id",
                    "type": "string"
                },
                "mic_opened": {
                    "description": "麦克风状态",
                    "type": "boolean"
                },
                "rtc_pull_url": {
                    "description": "连麦拉流地址",
                    "type": "string"
                },
                "user_avatar": {
                    "description": "用户头像",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户Id",
                    "type": "string"
                },
                "user_nick": {
                    "description": "用户Nick",
                    "type": "string"
                }
            }
        },
...
        "models.Status": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}